Get title to appear in document metadata
I noticed that the document was being shown by the PDF viewer as just
'git.pdf', which isn't very interesting.  It turned out that I'd put the
`\begin{document}` in the wrong location and needed to put this after
setting the title, author and date information.  Now the title appears
in the document metadata as expected.